requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
hubspot-api-client==7.5.0
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
